requirements.txt
# Actuales
fastapi
uvicorn
pydantic
python-dotenv
psutil
langchain
langchain-core
langchain-openai
langchain-chroma
chromadb
sse-starlette
python-multipart
httpx

# Nuevas dependencias para auth y caché
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
redis>=5.0.1
sqlalchemy>=2.0.0
alembic>=1.12.0
python-jose[cryptography]>=3.3.0